ADD:平台风险账户TOP10视图

+++ common-repository/src/main/java/cn/qkl/common/repository/mapper/AccountDynamicSqlSupport.java
@@ -41,7 +41,7 @@ public final class AccountDynamicSqlSupport {
     public static final SqlColumn<Date> updateTime = account.updateTime;
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    public static final SqlColumn<Integer> platformId = account.platformId;
+    public static final SqlColumn<Long> platformId = account.platformId;
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
     public static final SqlColumn<Integer> releaseNum = account.releaseNum;
@@ -68,7 +68,7 @@ public final class AccountDynamicSqlSupport {
 
         public final SqlColumn<Date> updateTime = column("update_time", JDBCType.TIMESTAMP);
 
-        public final SqlColumn<Integer> platformId = column("platform_id", JDBCType.INTEGER);
+        public final SqlColumn<Long> platformId = column("platform_id", JDBCType.BIGINT);
 
         public final SqlColumn<Integer> releaseNum = column("release_num", JDBCType.INTEGER);
