+++ common-repository/src/main/java/cn/qkl/common/repository/model/SocialPlatform.java
@@ -5,7 +5,7 @@ import javax.annotation.Generated;
 
 public class SocialPlatform {
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    private Integer id;
+    private Long id;
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
     private String name;
@@ -17,18 +17,18 @@ public class SocialPlatform {
     private BigDecimal socialPlatformIndex24h;
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    public SocialPlatform withId(Integer id) {
+    public SocialPlatform withId(Long id) {
         this.setId(id);
         return this;
     }
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
